index hell
Trying to make the main model have an optimized forward and backward algorithm, which is resulting in the worst indexing I've ever written. There is a 0% chance this works on the first try, and once it's done, I will pray I never have to figure out how it works again.