Started writing the final objects.
Now that all the algorithms are done, I'm starting to finalize parts of the library for initial release. The current objects are a huge mess and nowhere near working condition, but they're a decent start and hopefully will be done relatively soon. I wrote all of this with no wifi on a plane flight without knowing the syntax for subclasses, so I might have to make some pretty significant changes in the near future to fix some syntax issues.